+++ packages/EasyErrorHandler/src/Bridge/Laravel/Handler/Handler.php
@@ -190,7 +190,9 @@ class Handler implements ExceptionHandler
      */
     protected function isExtendedResponse(): bool
     {
-        return $this->config->get('easy-error-handler.use_extended_response', false) === true;
+        $extendedResponse = (bool)$this->config->get('easy-error-handler.use_extended_response', false);
+
+        return $extendedResponse === true;
     }
 
     /**
@@ -325,6 +327,9 @@ class Handler implements ExceptionHandler
      */
     private function translateMessage(string $message, array $parameters): string
     {
-        return $this->translator->trans(\trim($message), $parameters);
+        // TODO: rework after upgrading all the illuminate and laravel packages to ^6.0
+        $method = \method_exists($this->translator, 'trans') ? 'trans' : 'get';
+
+        return $this->translator->{$method}(\trim($message), $parameters);
     }
 }