+++ packages/EasyNotification/src/Interfaces/ConfigFinderInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EonX\EasyNotification\Interfaces;
+
+interface ConfigFinderInterface
+{
+    public function find(string $apiKey, string $providerExternalId): ConfigInterface;
+}