+++ packages/EasyRepository/src/Implementations/Doctrine/ORM/DoctrineOrmRepositoryTrait.php
@@ -6,6 +6,7 @@ namespace EonX\EasyRepository\Implementations\Doctrine\ORM;
 
 use Closure;
 use Doctrine\DBAL\DBALException;
+use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\ORMException;
 use Doctrine\ORM\QueryBuilder;
 use Throwable;
@@ -105,6 +106,11 @@ trait DoctrineOrmRepositoryTrait
     protected function createQueryBuilder(?string $alias = null, ?string $indexBy = null): QueryBuilder
     {
         return $this->repository->createQueryBuilder($alias ?? $this->getEntityAlias(), $indexBy);
+    }
+
+    protected function getClassMetadata(): ClassMetadata
+    {
+        return $this->manager->getClassMetadata($this->repository->getClassName());
     }
 
     protected function getEntityAlias(): string