+++ packages/EasySecurity/src/SecurityContextFactory.php
@@ -4,15 +4,14 @@ declare(strict_types=1);
 
 namespace EonX\EasySecurity;
 
-use EonX\EasySecurity\Interfaces\SecurityContextFactoryInterface;
 use EonX\EasySecurity\Interfaces\SecurityContextInterface;
 
 /**
  * Not final on purpose for BC compatibility until 3.0.
  */
-class SecurityContextFactory implements SecurityContextFactoryInterface
+class SecurityContextFactory extends AbstractSecurityContextFactory
 {
-    public function create(): SecurityContextInterface
+    protected function doCreate(): SecurityContextInterface
     {
         return new SecurityContext();
     }